Move encrypt functions to cipher
